RoomView: Do a server side and a local search if the scope isn't Room.
A search can be done for a single room or for all rooms. Since a
platform that supports event indexing will only index encrypted rooms we
need to do a local search as well as a server side search if we're
searching all rooms.

This patch adds the ability to correctly search all rooms.