MatrixChat: Don't await the decryption promise for live events.
Awaiting the decryption promise for live events seems to mess up the
event for Riot. If a room session key is already present the events seem
to be marked as unencrypted while if a session key is missing they never
get decrypted when one arrives.

To mitigate that wait for a Event.decrypted event instead and add the
event to the index when that event fires.